+++ src/renderer/components/match-view/MatchView.tsx
@@ -3,7 +3,8 @@ import fs from "fs";
 import path from "path";
 import ShareButton from "../misc/ShareButton";
 import ManaCost from "../misc/ManaCost";
-import { actionLogDir, ipcSend } from "../../rendererUtil";
+import { actionLogDir } from "../../rendererUtil";
+import { ipcSend } from "../../ipcSend";
 import DeckList from "../misc/DeckList";
 import RankIcon from "../misc/RankIcon";
 import db from "../../../shared/database-wrapper";
