+++ src/renderer/components/settings/SectionData.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { shell } from "electron";
 import Toggle from "../misc/Toggle";
 import Input from "../misc/Input";
-import { ipcSend } from "../../rendererUtil";
+import { ipcSend } from "../../ipcSend";
 import ReactSelect from "../../../shared/ReactSelect";
 import { parse, isValid } from "date-fns";
 import { useSelector } from "react-redux";
