+++ src/renderer/tabControl.tsx
@@ -15,7 +15,7 @@ import HomeTab from "./tabs/HomeTab";
 import MatchesTab from "./tabs/MatchesTab";
 import TimelineTab from "./tabs/TimelineTab";
 import OfflineSplash from "./components/main/OfflineSplash";
-import { ipcSend } from "./rendererUtil";
+import { ipcSend } from "./ipcSend";
 import SettingsTab from "./tabs/SettingsTab";
 import store from "../shared/redux/stores/rendererStore";
 import { reduxAction } from "../shared/redux/sharedRedux";
