+++ src/renderer/tabs/EventsTab.tsx
@@ -7,7 +7,8 @@ import EventsTable from "../components/events/EventsTable";
 import { EventStats, EventTableData } from "../components/events/types";
 import { isHidingArchived } from "../components/tables/filters";
 import { useAggregatorData } from "../components/tables/useAggregatorData";
-import { ipcSend, toggleArchived } from "../rendererUtil";
+import { toggleArchived } from "../rendererUtil";
+import { ipcSend } from "../ipcSend";
 import { getMatch, matchExists, eventsList } from "../../shared/store";
 import { reduxAction } from "../../shared/redux/sharedRedux";
 import store from "../../shared/redux/stores/rendererStore";
