+++ src/renderer/tabs/MatchesTab.tsx
@@ -9,7 +9,8 @@ import { MatchTableData } from "../components/matches/types";
 import { isHidingArchived } from "../components/tables/filters";
 import { TagCounts } from "../components/tables/types";
 import { useAggregatorData } from "../components/tables/useAggregatorData";
-import { ipcSend, toggleArchived } from "../rendererUtil";
+import { toggleArchived } from "../rendererUtil";
+import { ipcSend } from '../ipcSend';
 import { reduxAction } from "../../shared/redux/sharedRedux";
 import { matchesList, getMatch } from "../../shared/store";
 import store, { AppState } from "../../shared/redux/stores/rendererStore";
